Create new Venv for Python-based Project (#3289)
* add `interpreterPath` to create env options in python extension

* add checks for conda when setting env provider

Conda is not yet supported, so don't allow it as a python env provider when creating the new project.

* create a Venv and install `ipykernel` in it

Next: start the runtime in the Venv

* only pre-install `ipykernel` for existing python environments

* start the new env interpreter

* organize positronNewProjectService into code regions

* create the new blank files before env tasks

Env tasks can take a while, but creating the blank files is quick. This way, the blank files are interact-able while the env is set up in the background.

* log when we hydrate metadata in `doAutoStartRuntime`

* add user notifications and improve logging in new project service