Project Wizard: handle state changes to project directory (#3307)
* add state cleanup method to new project state

This allows us to clear out any state  properties that are irrelevant to the project being created, based on the project language.

This should fix issues with R projects not having their interpreter start due to waiting for python env steps that are irrelevant to R projects.

* fix wizard state persistence issues

* set projectNameFeedback in state

This allows us to avoid using a `useEffect` with an empty dependency array which requires a disabled linter warning.

* add missing early return

Signed-off-by: sharon <sharon-wang@users.noreply.github.com>

---------

Signed-off-by: sharon <sharon-wang@users.noreply.github.com>