Data Explorer: Add preliminary customizable float formatting options for data values, summary stats (#3310)
Partly addresses #2339 and #3210, and #3314. Adds a format_options parameter in the get_data_values and get_column_profiles backend methods to allow customized formatting of large, small, and medium-sized numbers. If the value threshold is above or below limits implied by the parameters, scientific notation is used. We've talked about dynamically trimming zero padding from the end of numbers (e.g. if all the floats displayed are integral, then we can trim the trailing zeros off all of them), but this will be a frontend-only change that can be done after.