Data Explorer: spec out "export_data_selection" backend method for exporting cells and cell/column/row ranges, and add Python pandas implementation (#3331)
This addresses #3313 and is a dependency of #2160 / #3288

This adds several selection OpenRPC types to describe single cell, cell range, row range, column range, and selections based on a list of column or row indices and allows the data to be exported to CSV, TSV (tab-separated, what Excel and Google Sheets need for pasting to work properly), and HTML.

There is currently no error checking to catch requests that may be too large/expensive -- probably easiest to do this on the front end for now.

pandas row indices are currently not being exported. I figure it's easiest to make this first step and then we can decide under which circumstances (e.g. only in the row-range/row-indices scenarios?) we want to include the row labels.