Data Explorer: Add column/row header and data cell context menus for data copying or other actions (#3374)
This PR contains a big push on adding column header, row header, and cell context menus to the Positron Data Explorer. Now that the context menus are in place, we can get feedback on them.

Note: I've renamed the anchor: HTMLElement parameter / property to anchorElement: HTMLElement because of the addition of a new anchorPoint parameter / property.