Enable project wizard in release builds (#3346)
* rename project wizard config flag to WIP

* add `wipFunctionalityEnabled` method to wizard state

* hide renv init when wizard WIP disabled

* hide conda env option when wizard WIP disabled

* fix interpreter feedback displaying while interpreters loading

* remove `isDevelopmentContext` check from project wizard wip setting

The project wizard WIP functionality will depend entirely on the setting being enabled, regardless of whether the IDE is a dev build or a release build.

* remove project wizard feature flag

* clean up positronNewProjectEnablement.ts

* update project wizard config setting with warning

* disable new project Create button when conda env selected

Creating the project with a conda env is not yet supported. Undefined behavior will occur if the user tries to create a project with conda as the env provider.

* remove extra args to `createWelcomePageLeft`