Create new Conda env for Python-based Project (#3385)
* support `conda` pythons in the project wizard dropdown

* support quickpick-less conda environment creation

Pass the `condaPythonVersion` to the `createEnvironment` command, so that a quickpick is not needed to determine the python version to use for the new environment.

* support creating and selecting new conda env

* make positron conda/venv handling more upstream-merge-friendly

* include conda env provider by default

* handle env providers loading state

* ensure the proper python path is set

Don't rely on lookup via validateMetadata later -- grab the interpreter
path right away after environment creation.

* clear the correct condaPythonVersion variable

* add runtime start request source to error message

* create new command `python.createEnvironmentAndRegister`

This commands allows us to create an environment and immediately register the interpreter with positron. Callers of this command will have access to the actual runtime metadata that was registered.

* fix interpreters loading check

* run prettier on python extension files

* remove unused `python.validateInterpreterPath` command

* use absolute path to env when installing ipykernel