Data Explorer: Add preliminary support for viewing Python polars data frames, disable most UI features that are not supported  (#3450)
Start of addressing #2185, will create sub-issues to break down work to implement all of the missing features. I have used the SupportedFeatures response to set disabled flags in the UI and added some CSS in consultation with Brian to toggle off row filtering, sorting, copy-and-pasting, expanding summary statistics. So basically we can just see the data values and the data types, along with the null percentages. We will be able to then incrementally add support for all of the missing features and communicate their support level via SupportedFeatures.

I added a couple of new items to SupportedFeatures for copy-paste and sorting, and added undefined checks so that the R kernel is not broken when this is merged.