Console Smoke Test: Added retries for stability (#3607)
### Intent
To resolve https://github.com/posit-dev/positron/issues/3536

There can occasionally be a race condition where Python starts up after
R and can steal focus of the console unexpectedly in the R test.

### Approach
Added a 1 retry, and changed the before block to `beforeEach` as only
the `beforeEach` will get re-run on a retry. So if python steals focus,
it will retry and run the `selectInterpreter` again, switching back to
the R console.

I also added the same retry to the Python tests as a precaution.

### QA Notes
Passed on Windows and MacOS