Switch from the faulty ActionBarTooltip component to a HoverService-b… (#3609)
### Intent

> Describe briefly what problem this pull request resolves, or what new
feature it introduces. Include screenshots of any new or altered UI.
Link to any GitHub issues that are related.

This PR does away with the faulty `ActionBarTooltip` component and
replaces it with a HoverService-backed tooltip manager.

Here's a screen recording:


https://github.com/posit-dev/positron/assets/853239/f7634621-f886-4308-82fb-28095f477a45

### Approach

> Describe the approach taken and the tradeoffs involved if non-obvious;
add an overview of the solution if it's complicated.

Notes:

- Just like every other use of HoverService, the time for immediate
display of a tooltip is 200ms. It'a also hard-coded, just like every
other use of HoverService. I would like to make this configurable
throughout Positron, just like `workbench.hover.delay`.
- There isn't a centralized `hoverLeaveTime` in the app. I think there
should be.

### QA Notes

> Add additional information for QA on how to validate the change,
paying special attention to the level of risk, adjacent areas that could
be affected by the change, and any important contextual information not
present in the linked issues.