use GitHub app to generate access tokens (#3606)
### Intent

This change removes all references to the `POSITRON_GITHUB_PAT` secret.
This secret was used in a lot of places to perform authenticated Github
actions; it is owned and maintained by one of the engineers on the dev
team (me).

### Approach

Instead of a PAT, we now use the new Positron Bot app's credentials by
asking it to issue an access token whenever one is needed.

Once all the jobs I've changed have been run successfully with this new
method, I'll remove the obsolete secret from the repository.

Here's a job run from this branch that demonstrates the credentials
being used successfully to fetch amalthea during the build (I canceled
it to conserve resources once confirming that access was successful):

https://github.com/posit-dev/positron/actions/runs/9602786991

### QA Notes

This change affects tests since they also use the PAT to access
repositories. I've updated the tests, too.