Streamline PR template and add CONTRIBUTING.md link (#3624)
Empirically, many of the sections of the current PR template are being
left empty. I wanted to propose a more minimal template that also points
to CONTRIBUTING.md (which needs to be worked on to add a new contributor
guide). It also puts the "instructions" in Markdown comments (so they
are hidden in the UI if the contributor does not delete them) to make
rendered PR descriptions easier to read. I also question whether the QA
notes section is needed, since QA is usually looking at issues when
doing verification rather than pull requests. So QA notes may be better
off there.

---------

Signed-off-by: Wes McKinney <wesm@apache.org>
Co-authored-by: Julia Silge <julia.silge@gmail.com>