Rename 'amalthea' repository to 'ark' repository (#3625)
### Intent

We need to rename the 'amalthea' repository to 'ark' in advance of
opening it up.

### Approach

This change performs the Positron side of the rename by updating build
scripts, dev scripts, and documentation as appropriate. It also removes
some cruft left over from the days when `amalthea` was nested inside
`positron-r` (those are the purely deleted lines).

Sequencing the rename will performed as follows:

- When this PR is approved, I'll perform the rename of the `amalthea`
repository to `ark`
- CI re-run against this PR to verify the changes (ensures we can
download the kernel from the new URL)
- Merge this PR to Positron

### QA Notes

This only changes how the R kernel is acquired, so it shouldn't have any
impact on product functionality -- as long as you can start R,
everything worked.