Data Explorer: Don't add commas when formatting for the data grid (#3611)
<!-- Thank you for submitting a pull request.
Please ensure that the code is up-to-date with the `main` branch.
-->

### Intent

> Describe briefly what problem this pull request resolves, or what new
feature it introduces. Include screenshots of any new or altered UI.
Link to any GitHub issues that are related.

Addresses https://github.com/posit-dev/positron/issues/3602 and
https://github.com/posit-dev/positron/issues/3603

### Approach

> Describe the approach taken and the tradeoffs involved if non-obvious;
add an overview of the solution if it's complicated.

### QA Notes

> Add additional information for QA on how to validate the change,
paying special attention to the level of risk, adjacent areas that could
be affected by the change, and any important contextual information not
present in the linked issues.