Add memory of last last size for PositronViewPane's for more dynamic minimum sizing (#3626)
Addresses #3620

### Intent

Add logic so that when the help pane is opened from a collapsed state it
will pop open to the previous size it was _unless_ that size sits below
some threshold (currently 100px). If it would pop open to a smaller size
then we go back to the `openFromCollapsedSize` if it exists. (Right now
half the window height for help/viewer).

All this logic is necessary because we allow the pane to get resized to
a zero height so that contents from the overlay webview don't spill out
and obscure other elements.

### Approach
Hook into the `layout()` method to remember the last size the panel was
set at and then use that value if it's above the min threshold,
otherwise fall back on the previous behavior of popping open to a larger
size to guarantee people can actually see the panels.

### QA Notes

You should be able to resize the help panel to a small (but not tiny)
size and upon collapsing and reopening it should stay how large you
expect it to be.

You should be able to resize the help panel to super short, then toggle
and on open the panel should pop open to a larger (aka visible) size.