Add ability to account for length of parent path in file validation (#3619)
Addresses #3561

### Intent

Take into account the parent of a path/ file name when deciding if it is
too long or not.
Previously we would make sure the file path directly was not over 256
characters, but we didnt account for the fact that we may be tacking
that onto the end of a parent path which would cause the path to go over
256 once combined.

### Approach

Add a new optional parameter for the file path validator function that
contains the parent path and use that to decide if the length of the
path is acceptable.

The entire path is passed rather than just the length so if in the
future we do something more complicated than taking the length we don't
need to change the api. Also is simpler to call.

### QA Notes
I tested by making the longest allowed combination of project name/
parent path and the new project was successfully created. I'm not sure
though if there's other more diabolical ways to mess with things.