Add layout service and full-sized part layouts (#3616)
Addresses  #3586 and #3577

### Intent

- Add a service that manages dispatching the positron custom layouts for
us.
- Add a few testing-related layouts like `"Maximized Sidebar Layout"`
that can configure positron layouts for easier smoke-testing.

Right now it is simple but provides a foundation in the future for more
advanced behavior like queing updates when workbench isn't ready or
making tweaks to existing layouts etc..

### Approach
Took the logic for the previous layout work that was spread across
methods on two existing services (`IWorkbenchLayoutService` and
`IViewDescriptorService`) and merged the details for how those methods
are called into an encapsulating positron layout service
(`IPositronLayoutService`).

At the same time, this provided a nice home for all the layout logic to
live so the previous scripts were moved into the new folder at
`src/vs/workbench/services/positronLayout/`

Also added three more layouts (not hooked up to quick-pick ui) for
maximizing specific parts of the UI: `workbench.action.fullSizedPanel`,
`workbench.action.fullSizedAuxiliaryBar`, and
`workbench.action.fullSizedSidebar`. These can be used to let the user
focus on a specific part of the page or for making automated testing
easier. (ala #3577)

### QA Notes

Existing layouts should work exactly as before. New layouts such as
`workbench.action.fullSizedAuxiliaryBar` should do what's printed on the
tin. Note that it's not possible for them to fully take up the entire
ide width or height because there's a minimum size protection for the
editor window.