Initial Positron extension API integration tests (#3615)
This PR adds a single rather basic test for the Positron extension API:
registering and then disposing a language runtime manager. (I noticed
that there may actually be a bug in the disposing part, see the TODO in
`runtime.test.ts`.)

We can decide whether we're happy with this approach before investing
too much in writing these tests.

The test will run as part of the existing `vscode-api-tests` extension
tests which are already a part of the `test-integration.sh` script. For
convenience, I've also added a `test-positron-api.sh` script that _only_
runs Positron API tests, so you should be able to do:

```sh
./scripts/test-positron-api.sh
```

from the root directory.

We'll likely need to fill out `TestLanguageRuntimeSession` as we add
more tests. I suspect it may end up looking a lot like the Zed and
JavaScript runtime session classes.

Co-authored-by: seem <mwlorgat@gmail.com>