Documentation added to smoke test README and code (#3662)
<!-- Thank you for submitting a pull request.
If this is your first pull request you can find information about
contributing here:
  * https://github.com/posit-dev/positron/blob/main/CONTRIBUTING.md

We recommend synchronizing your branch with the latest changes in the
main branch by either pulling or rebasing.
-->


Just a pass on adding simple documentation. For the readme, the changes
focus on describing the default project and debugging in CI. For the
code, the documentation primarily identifies the overall purpose of each
class.

### QA Notes

Just documentation.  Run status should not change.