Add smoke test for help panel resize logic (#3639)
Addresses #3623 

## What?
Adds a smoke test that focuses on the help panel resizing behavior. 

## How?
Tests the workflow of opening an empty help pane making sure it opens
and can be resized and remembers the new size.

This PR also adds a method on the playwright driver: 
- `.clickAndDrag()` which simulates clicking and dragging the mouse from
a given location. Used to test resizing with the sash in help test

It also adds a few methods on the existing `PositronHelp` class
- `.getHelpContainer()` which grabs the entire help panel's locator.
- `.getHelpHeader()` which gets the locator for the bar dividing the
help pane when it's docked with other views in a view container.
- `.resizeHelpPanel()` which uses the above `.clickAndDrag()` method to
resize a docked help-view by dragging the resize sash some delta.

This also adds a fix for a bug found via these tests! That being that
our remember-resize-size logic was not accounting for the height of the
header panel. This is now taken into account.

### QA Notes

There's maybe a bit more logic within the test itself than other smoke
tests. I'm open to reconsidering this structure!

I noticed weird inconsistency with sizing when running locally. Not sure
how much of a pain this will be on CI.

---------

Signed-off-by: Nick Strayer <nick.strayer@posit.co>
Co-authored-by: Jon Vanausdeln <jonv@live.com>