Data Explorer: Implement export_data_selection for polars to enable copy-and-paste (#3818)
Addresses #3454. I also added a `supported_formats` field in
`SupportedFeatures.export_data_selection` that will have to be added to
Ark in the next iteration. It isn't currently used anywhere in the UI so
its absence will not affect anything for now.

Note that polars does not seem to have HTML export out of the box so
only CSV and TSV is supported for now.

### QA Notes

Try copy-and-pasting single cells, rectangular ranges, whole columns or
rows, or a random selection of columns or rows.