Upload Smoke results to TestRail (#3831)
### Intent

Addresses #3821 and provides for Automated Smoke test results to be
uploaded into Testrail

### Approach

The TestRail API will parse the junit xml results file and send results
for each test with a TestRail ID in the name. The tests Run is
immediately closed after results are added. Currently, there is no
milestone associated, but we could add one later if needed (perhaps when
we start using release branches, etc)

* Added TestCase ID to each test `it` block
* Changed the Junit results file name to be consistent (results.xml)
* Updated the workflow yml file
  - pip install the TestRail CLI package
- Added step to generate at Test Run title using the current Date. This
is linux specific for now. When we add more platforms, this will need to
be update for the new platforms.
  - Added step to upload the results via the TestRail CLI

Note, the currently uses an API key for my TestRail account. We don't
have a service account at this time, but could in the future and switch
to that account.

### QA Notes
Test on full smoke run, see:
Full Test Workflow Run:
https://github.com/posit-dev/positron/actions/runs/9771103109
TestRail Test Run: https://posit.testrail.io/index.php?/runs/view/2832

It won't do anything for locally run tests as the upload step is only in
the GitHub Action workflow. You an run that command locally if you want
to upload your locally run results (you'll need your own TestRail API
Key)