Smoke test: copy and paste from the console (#3774)
This is a smoke test for copying from the console and pasting to the
console that also performs some error checking.

It first runs some bad code in the console (undefined variable) and
awaits the expected error message. It then clicks the console and
selects all its contents using CTRL/CMD-A CTRL/CMD-C and evaluates what
is stored in the clipboard looking for the same error. Then it pastes
the full contents of the clipboard into the console and awaits a
different error based on the pasted data not being valid code.

It performs these steps for Python and then for R.

### QA Notes

All smoke tests should pass.