Fix leaking commands from the connections extension to the command pallete (#3853)
This PR makes a few commands that are internal to the connections pane
extension to not appear in the command pallete. Which addresses #3841

Also adds command to expand connections, which will make testing easier.
We shouldn' need to rely on specifics of the TreeView API for smoke
testing the connections pane. Addresses #3840