Cache cell parsing in `positron-code-cells`
Prior to this PR, the parser would be called for almost any interaction with
the text (e.g. moving the cursor around). To fix this, the `cellManager` logic
was moved to a `documentManager`. There is an array initialized during initailization
that stores documents, `documentManagers`. When `TextDocument`s are loaded or closed
the array is updated. Then, parsed cells can be obtained via a document's `uri`.