Data Explorer: Add "search_schema" support for polars, add 'match data types' support, introduce "column filter" RPC type, refactor row filters a bit (#3904)
This adds a `match_data_types` filter type in the `search_schema` RPC to
support column selection and filtering workflows. This is
implemented/tested for both pandas and polars -- no user-visible
capabilities in this change.

I also refactored the row filter RPC types a bit to support
extensibility and code reuse. We will need to coordinate with Ark when
merging this.

Related Ark PR: https://github.com/posit-dev/ark/pull/424