Add method for updating size of the window for smoke tests (#3943)
### QA Notes

I only added one usage of this now, but perhaps in the future, we should
start enforcing a standard size for our smoke tests to have consistent
results.

You can check that it's "working" by watching the app as the resizing is
happening. (Use debug stepping if it happens too fast!) Note that it
won't actually resize the app window, but will simply clip. In a
headless instance, it will behave as expected.