Use just-in-time access token for checking CLA (#3985)
This is a speculative fix intended to address the frequent 403 access
errors we're seeing with the CLA assistant. Instead of using a PAT
provided as a secret to the Github Action, we generate an access token
intended for one-time use and supply that to the upstream action.

### QA Notes

N/A, build change only.