Add method for hiding views and use to hide jupyter tab (#3977)
Addresses #3901

This PR hooks into the view descriptor service and adds a way to
register views we don't want to show up in the UI. This is useful for
instances where an extension we don't control contributes the view but
our UI duplicates the functionality.

The specific instance implemented here is to hide the jupyter tab that
was showing up even though our own variables tab replicates the behavior
(in a superior manner if I do say so).

## Implementation notes
This is done in the view descriptor service with the goal of having the
smallest blast radius. Basically, the view registration is allowed to
run to completion and it's just never added to the UI. There's always a
chance though that some logic within the view's UI could be important,
but I doubt it. Also, we already mess with the view descriptor service
for the layouts so we can keep the modification of non-positron scripts
to a minimum.

We _could_ do this by utilizing the dynamically created
`${viewDescriptor.id}.removeView` command that's used for the
right-click -> hide interface, but it's not clear if that could be
accomplished without flashing the non-desired view for a second (also
we'd have to add logic for detecting the registration of the view.)

### QA Notes

Everything should operate exactly as before with the exception of, when
a jupyter notebook is open (the original VSCode notebooks, not the
positron notebooks) the jupyter tab should never appear.