Address 403 errors in CLA validation job (Take 2) (#3994)
This is a second attempt at addressing the sporadic 403 errors emitted
by the CLA assistant.

It turns out that the kind of reusable workflow exposed by the assistant
is not callable from a job step (i.e. it's not a composite step). So in
order to get a secret into it, we need create the secret from a separate
job, promote it to an output, and _then_ we can use it in the reusable
workflow.

### QA Notes

No QA needed, CI change only.