Share positron-icon.svg between UI elements (#3997)
The positron icon as an SVG resource is needed in additional places
beyond the welcome page. This change moves it to a more central location
for UI media and updates references to ensure the Positron icon is
displayed instead of a generic default.

Addresses #3996


### QA Notes

While Mac does not render a title bar icon for the product, you should
see the Positron logo on the Welcome Screen editor tab icon. Windows
should, however, now show the Positron icon in the top left on the title
bar. Additionally, the positron logo should still render in the body of
the welcome editor too on all platforms.