Run show-version scripts in node explicitly (#4039)
This change fixes an annoying issue in Windows dev environments that can
cause Visual Studio to open during the incremental build.

The problem is that Gulp is invoking `show-version.js`. On Unix-alikes,
the shebang comment in this file causes it to run under Node, but on
Windows, invoking the file opens it in the default editor for `.js`
files.

The fix is to use Node explicitly instead of relying on the shebang. 

### QA Notes

N/A, build change only.