Console input smoke test - enter your name (#4022)
This is a simple console user input test. In both Python and R, the
"user" is prompted to enter their name and "Hello ${name}" is echoed
back. The test waits until the echo is received in the console output to
determine pass/fail. This PR also adds in a pasteToConsole function that
simulates a user pasting code as opposed to entering code via a command
(to use the console itself more).

### QA Notes

All smoke tests should pass!