Data Explorer: work around errors caused by upgrading from polars 0.20.x to 1.2.x (#4105)
polars 1.2.x contains various behavior changes that had to be worked
around, causing errors like #4070. I reported some of the issues as bugs
upstream in polars (they may or may not be considered bugs in polars --
they look like bugs to me or at minimum unpleasant sharp edges). I was
able to get the unit tests working on both 0.20.x and 1.2.x but we have
a blind spot in our CI right now in that we do not test > polars 1.0
yet.

### QA Notes

These failures arose because the polars version in the smoke test CI
setup wasn't pinned -- it would have been sufficient to run the unit
test suite for both 0.20.x and 1.2.x to find these issues.