Add non-overlay webview creation to notebook webview service and use for rendering html in positron notebooks (#4078)
This PR has three main parts to it as told by the (first) three commits:

1. Update the `IPositronNotebookWebviewService` to expose the
`.createRawHtmlOutput()` method and allow that method to create standard
`WebviewElement`s in addition to `OverlayWebview`s. (Also adds the
ability to create the webview without an active runtime due to runtime
requirements preventing notebooks from showing html content until the
kernel is connected.)
2. Add knowledge of the current language runtime attached to a notebook
to the `PositronNotebookInstance` class. (Also uses that knowledge to
make the kernel badge more accurate)
3. Replace the existing rendering of HTML content in notebooks with the
`IPositronNotebookWebviewService.createRawHtmlOutput()` and all the
goodies that we get (and will get in the future) from that.

### QA Notes

This should be a simple under-the-hood swap; nothing should look or work
differently.