Data Explorer: Add formatting option to truncate large strings, default to 1000 characters, implement for pandas/polars (#4121)
Addresses #3823. This adds `FormatOptions.max_value_length` to truncate
stringified values at some number of characters. There's currently no
indication that the value has been truncated (but getting to the end of
a 1000-character string in the UI would be hard). If you copy-paste a
cell, you will get the entire value, not the truncated version.

### QA Notes

Tricky to test -- I temporarily modified the TypeScript code to use some
small number (like 32) to inspect that things are truncating correctly.
If you load a data frame with a lot of large string values (think like
1-megabyte strings), then the UI should remain responsive. Here's an
example that would put a hurt on the current version of Positron:

```
x = 'aahrealmonsters' * 100000
chunky_df = pd.DataFrame({'a': [x] * 1000})
```