Dont try and show viewer-widget output in the console (#4140)
Addresses #2079


Currently when a runtime message is sent out with the type
`RuntimeOutputKind.ViewerWidget` it is both captured by the viewer pane
_and_ the console. This manifests in some confusing behavior as seen
when trying to print out a great tables table. In this case the html for
the table is rendered as one would hope in the viewer pane but also the
`text/plain` data is also shown in the console. This can cause some very
polluted console output.

This PR adds a condition to the positron console service's message
handler callback that exits early if the message received is of the
viewer-widget type.
<img width="600" alt="Screenshot 2024-07-24 at 11 38 07 AM"
src="https://github.com/user-attachments/assets/13729c99-a718-4eb3-ba7d-9558b0d2ae73">

## Potential changes
- Right now the console simply shows nothing. This is clean but maybe is
confusing for the user? We could alternatively show something like
`<output in viewer pane>` or the like.
- There may be a more appropriate place to do the filtering of message
destinations other than the respective subscriber callbacks.

<!-- Thank you for submitting a pull request.
If this is your first pull request you can find information about
contributing here:
  * https://github.com/posit-dev/positron/blob/main/CONTRIBUTING.md

We recommend synchronizing your branch with the latest changes in the
main branch by either pulling or rebasing.
-->

<!--
  Describe briefly what problem this pull request resolves, or what
  new feature it introduces. Include screenshots of any new or altered
  UI. Link to any GitHub issues but avoid "magic" keywords that will 
  automatically close the issue. If there are any details about your 
approach that are unintuitive or you want to draw attention to, please
  describe them here.
-->



### QA Notes
Printing a great tables table (or any other viewer-widget type outputs)
should not also show output in the console.