Add OpenRPC method for evaluating a `when` clause (#4165)
Goes along with https://github.com/posit-dev/ark/pull/449

I was thinking about how to address
https://github.com/posit-dev/positron/issues/2697 and have a proposed
approach in this PR plus the accompanying PR for ark. I looked at how
various extensions (like the git extension, etc) check for whether we
are in a git repo in the workspace and all that checking is typically
based on context keys. In an extension specifically, you can check those
keys with a ["when"
clause](https://code.visualstudio.com/api/references/when-clause-contexts)
and I realized that approach (i.e. a string) could work well as an
OpenRPC method and would be flexible for other context keys we will need
to check when we are not in the main thread.

Alternatively, we could add methods to return more specific info (are we
in a git repo? etc etc etc) but I like the idea of using "when" clauses
for this, like an extension could directly for commands, etc.

### QA Notes

Evaluating code like this in R should show the correct results for your
situation:

```r
.ps.ui.evaluateWhenClause("isLinux || isWindows")
.ps.ui.evaluateWhenClause("isMac")
.ps.ui.evaluateWhenClause("gitOpenRepositoryCount >= 1")
```

---------

Signed-off-by: Julia Silge <julia.silge@gmail.com>