Data Explorer: Add schema cache for small schemas to facilitate checking for inplace updates with pandas and polars (#4184)
This addresses #4144 and #4145 as much as possible. I added a "small
schema cache" for the scenarios where there are 100 or fewer columns and
it isn't too expensive to eagerly compute all the data types for pandas
and polars. For very large data frames, inferring data types for object
columns can have arbitrarily large cost and so in these instances, we
need to continue to send "schema updated" events after code cell
executions to avoid staleness in the UI.

This approach can continue to be refined (e.g. maintaining a partial
schema cache for the columns that have been computed because of
`get_schema` requests) but working on this logic is rather tedious, so
let's consider this to continue to be a work in progress.

### QA Notes

Nothing in particular to verify here except that the smoke tests and
other workflows continue to function as expected.