Capture screenshots when fixtures fail; handle "detecting kernels" during notebook open (#4193)
Two fixes:
* Screenshot the application if the Python or R startup fixture fails.
This has happened recently in cases where the test cannot detect
"started" in the console to know when to move forward. We need to get a
look at the full app when this happens to debug further.
* Allow the notebook startup process to show "Detecting kernels" in
kernel selector. This has rarely happened, but we should be able to
handle it by waiting for "Select Kernel" or the valid kernel name when
"Detecting kernels" shows up. Previously we only expected a valid kernel
or "Select Kernel"

### QA Notes

All smoke tests should pass. If a startup fixture randomly fails to
detect "started" kernel, screenshot the application before exit.